requirements.txt
# Core dependencies
gfootball>=2.10
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# RL libraries
stable-baselines3>=2.0.0

# Logging and monitoring
tensorboard>=2.14.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional
psutil>=5.9.0
